requirements.txt
# requirements.txt
langchain==0.2.1
langchain-openai==0.1.7
python-dotenv==1.0.1
tenacity==8.2.3
rich==13.7.1
PyYAML==6.0.2

